Shortest Job First (SJF) Algorithm in C
Shortest job first scheduling is the job or process scheduling algorithm that follows the non pre-emptive scheduling discipline. In this, scheduler selects the process from the waiting queue with the least completion time and allocate the CPU to that job or process. Shortest Job First is more desirable than FIFO algorithm because SJF is more optimal as it reduces average wait time which will increase the throughput.